+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [3.0.4]
+- Select languages in user language rule
+- Fix position check in org unit user type rule
+
 ## [3.0.3]
 - Fix move request in create course action